Subrpositoris will be replaced by setuptools package dependency.
The ony thing subrepos was used is the ability to develop library and
use it as dependency same time. It happens that "python setup.py develop"
command is intended exactly for this purpose, so subrepos are dropped - they
prevents sharing code via "hg serve" without internet connection. And such
setups are common in corporate environment.